requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
torch==1.10.0+cpu
torchvision==0.11.1+cpu
requests==2.26.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
opencv-python==4.5.3.56
pillow==8.4.0
tqdm==4.62.3
streamlit==1.3.0
pickle-mixin==1.0.2
zipfile36==0.1.3
dlib==19.22.1
